speedtest/helper: store Upload and Download values in bytes per second.
This is a change that violates backward compatibility.
Anyway storing Mpbs rounded to 3 decimals can lead to lose information for slow connections.

+++ report/csv.go
@@ -11,8 +11,10 @@ type CSVReport struct {
 	Address   string    `csv:"Address"`
 	Ping      float64   `csv:"Ping"`
 	Jitter    float64   `csv:"Jitter"`
-	Download  float64   `csv:"Download"`
-	Upload    float64   `csv:"Upload"`
-	Share     string    `csv:"Share"`
-	IP        string    `csv:"IP"`
+	// Download bandwidth in bytes per second
+	Download int64 `csv:"Download"`
+	// Upload bandwidth in bytes per second
+	Upload int64  `csv:"Upload"`
+	Share  string `csv:"Share"`
+	IP     string `csv:"IP"`
 }
